Fix KeepLogSeg() unittest
The current WAL record pointer passed by the test to KeepLogSeg was
older that _logSegNo.  This cannot happen, so fix it by increasing it
beyond _logSegNo.  The test needs to be refactored heavily to make use
of new WAL format, but let's do it as a follow up.

Reviewed by: Ashwin Agrawal <aashwin@vmware.com>
Reviewed by: Paul Guo <paulguo@gmail.com>

(cherry picked from commit e4ec851b151267fe3209af6804e438611836a875)