recoverseg/addmirrors/movemirrors: Revert catalog update for failures
Currently, for non inplace full recoveries, we need to modify the
gp_segment_configuration table to reflect the modified value of
host/port/datadir. Because of the new workflow(see previous commits),
where we do this update before running the recovery process, it's
possible that any failures during the recovey process can leave the
catalog in an unexpected state for subsequent recoveries.
Note that in place full recoveries don't modify the catalog at all.

This commit modifies the code to create backout sqls, which revert any
updates made to the gp_segment_configuration table. We also add an entry
to the gp_configuration_history table for each dbid that gets reverted
to it's original configuration. We only run the revert sqls for the
dbids that failed during their basebackup operation.

Also note that if for a mirror, the basebackup operation completes
successfully but the start operation fails, then we don't revert the
catalog update.

Special case for gpaddmirrors
------------------------------
Note that for gpaddmirrors, we decided not to revert the catalog updates
in case any of the mirrors fail during the basebackup operation. This is
because the revert sql for gpaddmirrors would remove that mirror's entry
from gp_segment_configuration. Since gpdb only supports "all primaries
have mirrors" or "no primaries have mirrors", reverting catalog for only
a subset of mirrors won't be possible.

So in order to support the revert functionality, we would have to remove
all the mirrors from gp_segment_configuration(even ones that succeded).
This would mean that the user would have to unnecessarily run the
basebackup operation for all those successful mirrors again. Instead if
we don't revert, the user can fix the underlying problem and run
gprecoverseg instead, which will be much faster since we would only have
to recover the mirrors that failed previously.

Co-authored-by: Nikhil Kak <nkak@vmware.com>

(cherry picked from commit 5daf37f08e82ab4d5013ae77e59025b8db8b9926)