recoverseg/addmirrors/movemirrors: Update catalog before recovery
A previous commit added code to start the segment as soon as it has
successfully recovered. For the segment to be marked as 'up', we also
need to make sure that any catalog updates(if any) are done before
trying to start the segment. This commit updates the catalog before
attempting to recover.

If any of the segments fails during the recovery process, we would also
have to make sure that we undo any catalog changes for that segment.
This will be taken care of in a future commit.

(cherry picked from commit b557f47eabd8d01230f0ff038c052ee8d731bc93)