Fix bug in gpconfigurenewsegment
A previous commit
https://github.com/greenplum-db/gpdb/commit/ed5a10a862b5ea9fddf0c7adce800f89e38f8277#diff-0d46beb91a7c4fcc4df30e273b0f1e3496cd2df1bcd93e3e879f36ae1cf9ac95
had changed the argument names of the pg_basebackup function but missed
changing it in the gpconfigurenewsegment file.

Note that because of the above commit, this code path isn't called
anymore but it's good to fix the bug anyway.

Co-authored-by: Divyesh Vanjare <vanjared@vmware.com>
(cherry picked from commit 948eea101f76f582afda7838bc986cf14318ced2)