Restore visibility of some GUCs to be viewable from pg_settings
The checkpoint_timeout, archive_command, and data_directory GUC
settings were hidden back in the early GPDB 4.0.X days to prevent
users from messing with them. It should be fine to realign the
visibility of these GUCs with how they are in Postgres
(checkpoint_timeout and archive_command visible to all users and
data_directory visible to superusers or users with
pg_read_all_settings permission). They won't be documented in
Greenplum GUC documentation so they'll still be somewhat hidden to
individual users but not for us developers and the tools we build that
may require checking their GUC values. This is mainly needed now to
allow external tools such as pgbackrest to work properly.