resolve update distributed key with oids fixme (#13343)
In the older version we can create table with (oids=true);
therefore, each row has an OID, when we update distributed columns of a table,
a row may move from one segment to another one. so we must ensure that the
target list contains the old OID so that the Split Update can copy it to the new tuple.

but the newest version of GPDB does not support create table with (oids=true).
when merging postgres12, the logic of codes is removed by this commit:
https://github.com/greenplum-db/gpdb-postgres-merge/commit/7c3bd14c32a746c8f8a3e59898d8e587638054fb
But the comments remain. Now we remove related comments also.

we just modify comments, so there is no test case.

(cherry picked from commit bbc8cc30985730961ba0c65c4bda13119b09022d)