gprecoverseg rebalance is failed with timeout while promoting mirrors
Currently gprecoverseg -r waits for a maximum of 30 seconds for preferred primaries to be promoted. This may not be enough and results in an exception

gprecoverseg rebalance workflow
1. All the segments are up and running but the roles are switched
1. User runs gprecoverseg -r
1. gprecoverseg shuts down all the acting primaries(preferred mirrors)
1. gprecoverseg triggers fts probe
1. gprecoverseg sends a test query (CREATE TEMP TABLE) to start a distributed transaction to test if all the preferred primaries were promoted
1. gprecoverseg waits for a max of 30 secs and checks if the CREATE TABLE query succeeded, else raises timeout exception
1. gprecoverseg starts the stopped preferred mirrors

The problem is that in some cases, the preferred primary may not be promoted in 30 seconds which results in an exception. For now we have decided to increase the timeout to 10 mins which should be enough for most use cases.
Eventually we will revisit this logic and make it a bit more robust.

(cherry picked from commit d649959b1f3bf4e19da0dc5a46f5e6aa97ef6cbf)