Revert "Remove unused Dynamic (bitmap/index/table) Scan code"
This reverts commit 806fc86bc12d181b3f870da9e393296bc4e01030.

Previously, we added support in ORCA for the postgres 10 style of
partitioning and utilized the postgres seq/index/bitmap executor nodes.
However, for plans with many partitions and many columns, this resulted
in large plan sizes (on the order of 100MB+). When combined with many
slices, this would cause increased memory utilization, sometimes going
OOM.

This series of commits reverts to using the Dynamic executor scan nodes.
We are keeping the logic to read from the GPDB7 catalog and the internal
Orca changes to handle static partition elimination within Orca. The
only intention of these changes is to use Dynamic executor nodes and
reduce the plan size. The plan shapes are the same.

For some numbers, we found that for tables with 1k+ partitions and 1k+
columns, the dispatch plan size for a single `select * from foo` was
100MB vs 300KB with this new approach. By itself, 100MB may not seem
like a lot, but if there are multiple partitioned tables with multiple
joins and slices, this can quickly become quite a bit.

(cherry picked from commit 38c560f90bcabcf11bba67a5627737ae7adae5f8)