Validate sql exec_location in CreateFunction (#13431)
GPDB add support for CREATE FUNCTION EXECUTE ON [COORDINATOR | ALL SEGMENTS] in #3312. According to
the commit message and Greenplum Doc, this CREATE FUNCTION options are only supported for
set-returning functions, but we didn't validate this limitation in `CreateFunction()` which 
`AlterFunction()` does.
This leads us to have several built-in functions and functions for test cases violate this limitation.
Some of them are not set-returning functions but have EXECUTE ON SEGMENTS attribute. This would return
multiple rows, one for each segments.

For example:
```
postgres=# select proname, proretset, provolatile, proexeclocation from pg_proc where proretset != 't' and proexeclocation != 'a';
             proname              | proretset | provolatile | proexeclocation
----------------------------------+-----------+-------------+-----------------
 brin_summarize_range             | f         | v           | s
 gin_clean_pending_list           | f         | v           | s
 gp_wait_parallel_retrieve_cursor | f         | v           | c
 gp_request_fts_probe_scan        | f         | v           | c
 brin_summarize_new_values        | f         | v           | c
(5 rows)

SELECT brin_summarize_range('brin_summarize_idx', 4294967295);
 brin_summarize_range
----------------------
                    0
                    0
                    1
(3 rows)
```

In most cases, functions with the EXECUTE ON attributes are used to perform queries on
distributed data or to execute functions on all segments. It is reasonable for these
functions to return set. There are also cases where the functions need to be executed
on the coordinator but does not need to return a set,  creating these functions as
set-returning function and using it like `SELECT func_exec_on_coordinator();` will not
affect the execution result.

This commit added validation for this limitation in `CreateFunction()` and modified
several returning non-set built-in functions which execute on master or all segments.