Revert PG_TRY() / PG_CATCH() surrounding of CFI. (#13567)
A history commit ee255b5 introduced a `PG_TRY() / PG_CATCH()` surrounding to
`CHECK_FOR_INTERRUPTS`, which seems not necessary and may cause some trouble,
according the discussions here:

* This TRY-CATCH is not enough of a guard, what if there are other avenues for
indirect CFI calls outside this `TRY-CATCH` and before signalQEs is called?
@kainwen mentioned that memory allocations can also trigger a CFI. So we will
have to reevaluate this fix to wrap the CFI call and rethrow at the end of the
function.

* @ashwinstar mentioned `cdbdisp_cleanupDispatcherHandle` will be called
during either `AtAbort_DispatcherState` or `AtSubAbort_DispatcherState`, so the
tricky specialized handling for the` cancelRequested` before may not be
necessary.

* There is also a `CdbDispatchHandleError()` call in `mppExecutorCleanup()`,
which is inside the `PG_CATCH` inside `PortalRun()`.

So it's better to revert this change to avoid nested TRY-CATCH now.