Fix missing Redistribute on top of Split Update with Orca
When deriving distribution spec for `CPhysicalSplit`, ORCA checks if its
disjoint with the hashed colrefs aswell as the equivalent hashed colrefs
(PdshashedEquiv()). For certain nested queries with a JOIN on the same
column, it is possible that the equivalent hashed colrefs also have
other equivalent hashed colrefs. ORCA did not consider such cases and
would skip recursively checking for hashed-equivalent colrefs and
include it to check if its disjoint with the colrefs being modified.
Thus generating a plan without a Redistribute on top of a Split node
when updating the distribution column. This would lead to wrong results.
This commit fixes the issue.

Co-authored-by: gpopt <103469259+gpopt@users.noreply.github.com>

Close  #13416

(cherry picked from commit 86118fe28cbb0be2d77c29db9080858b6696c2b0)