Fix a flakiness in test recovery test overwrite_contrecord
The test fails on some pipelines because there's a checkpoint WAL
being written after we write a new segment. So when the standby
is created from the same backup directory, it will complain about
not finding the checkpoint WAL, since we deleted it in the test.

It's flaky in some pipeline but not others because if the checkpoint
creation is slow we'll avoid the failure.

The reason which it writes the checkpoint WAL is because TAP test
set max_wal_size to 128 when creating the node. So the calculation
for how many segments to create checkpoint becomes:
(see CalculateCheckpointSegments)
    floor(128 / 64 / 1.5) = 1
So basically it'll create checkpoint for every new segment, which
is not good.

Also part of the reason is that GPDB has segment size = 64MB. In
upstream it's 16MB so the above calculation will get 5 which is ok.

Since the reason that the PG sets it to 128 seems to due to saving
space consumption, let's consider still use a smaller value but
increase it to 512MB which corresponds to the ratio of GPDB segment
size over PG segment size.

(cherry picked from commit bae05aa8b87e954e68d24bff759973aa6c343c12)