Support intermediate aggs in Orca plans (#13707)
Orca in GPDB6 has support for intermediate aggs, which isn't used in postgres.
This is useful when we have a DQA and a regular "ride-along" agg.
However, we need to differentiate when we should run the
combine/final/trans functions when this ride-along agg is present.

This commit re-adds support for intermediate aggs. The logic here is the
same as 6X, however, instead of explicitly using the aggstage, we use
the aggsplit, which is determined from the aggstage. The logic is
defined in `AGGSPLIT_INTERNMEDIATE`.

The changes in nodeAgg.c are to allow the aggref and aggstate to differ
for an aggregate. This is necessary and expected in the case of an
intermediate agg, as the loop will iterate over each aggstate->aggs, but
the aggsplit can now be different between the aggref and the aggstate.
Thus the aggsplit references are also changed to use aggref instead of
aggstate.

(cherry picked from commit 60179b110bb8a46c7ae01e1b204c4159c47e8173)