Just use default on commit behavior to child tables when we create partition table (#13687)
When we create a temporary table, we can use ON COMMIT to control the behavior
of temporary tables at the end of a transaction block, such as DELETE ROWS, means
delete all rows in this temporary table after this transaction is committed.

But when used on a partitioned table, this is not cascaded to its partitions, so we can just
use ONCOMMIT_NOOP (default behavior) when we create child partition tables.

And If we use the parent on commit action, it will lead to inconsistent behavior between
Greenplum and PostgreSQL, which we need to avoid.