Add writing CHECKPOINT to gpactivatestandby
Commit 1fe901d removed the rety loop for checking if promotion is
completed for standby. As side effect it also deleted the logic to
force CHECKPOINT after promotion. Bring that logic back as it required
to write CHECKPOINT after pg_ctl promote mainly to satisfy
tests. pg_ctl promote uses fast_promotion and hence pg_controlfile
gets updated with status as DB_IN_PRODUCTION but given CHECKPOINT is
not forced the TLI value is not updated. gpstart uses the TLI value to
validate if standby is promoted or not. Hence, bringing back the logic
to force CHECKPOINT after promotion.