gpcheckcat needs to reset GV.retcode before starting to check a database (#13735)
* Reset GV.retcode before running test for each database using gpcheckcat

Previously while running gpcheckcat, if the check fails on one database then the value of GV.retcode was not reset before running the checks on the next database in the list.Due to this a check failed on one database was causing a AssertionError for the next database in the list which passes. With this PR added a change for resetting the value of GV.retcode after each database check.