Remove palloc() call in ResWaitOnLock (#13701)
If palloc() failed in ResWaitOnLock due to OOM, it will trigger
a PANIC "Waiting on lock already held" because partitionLock is
not released correctly.

The first attempt is to add a try/catch surrounding for this
palloc() call, and release the partitionLock before rethrowing
error. But it is not good as ResIncrementRemove is not called.

So use local char new_status[160], length 160 is enough.

----

Steps to reproduce:

Add a fault at the palloc() position:
    SIMPLE_FAULT_INJECTOR("reslock_acquire_before_wait");

```sql
-- superuser:

-- test user session 1:
> SELECT pg_sleep(120);

-- test user session 2:
> SELECT version();
PANIC:  Waiting on lock already held! (lwlock.c:668)
server closed the connection unexpectedly
```

(cherry picked from commit 73da0f32f3860cb9ca128221761bc49929586deb)