[ORCA] Fix duplicate stats reset (#13817)
Following scenario led to crash due to missing statistics:
  ```sql
  CREATE TABLE t1 (c11 varchar, c12 numeric(15,4));
  CREATE TABLE t2 (c2 varchar);
  CREATE TABLE t3 (c3 varchar);

  SET allow_system_table_mods=true;

  UPDATE pg_class SET relpages = 97399::int, reltuples = 9106730.0::real, relallvisible = 0::int WHERE relname = 't1';
  UPDATE pg_class SET relpages = 68553::int, reltuples = 7054520.0::real, relallvisible = 0::int WHERE relname = 't2';

  SET optimizer_join_order=exhaustive;
  SELECT
       (SELECT c11 FROM t1) AS column1,
       (SELECT sum(c12)
          FROM t1
                  INNER JOIN t2 ON c11 = c2
                  INNER JOIN t3 ON c2 = c3
                  INNER JOIN t3 a1 ON a1.c3 = a2.c3
                  LEFT OUTER JOIN t3 a3 ON a1.c3 = a3.c3
                  LEFT OUTER JOIN t3 a4 ON a1.c3 = a4.c3
        ) AS column2
  FROM t3 a2;
  ```

Underlying cause is due to the fact that derive and reset for group
stats was not symmetric. In the case of "exhaustive", multiple xforms
may be run on the same group that derive stats before applying and reset
stats afterward. Prior to this commit it was possible to have a group
with "dirty" stats where child nodes may have been cleaned up, but the
group still tecnically has stats object. If it is a duplicate of another
group then it was possible to "trick" the other group into believing
that the stats were already derived. That fake news could lead to a
crash.

Co-authored-by: Jingyu Wang <wjingyu@vmware.com>
(cherry picked from commit b1d2e27ef858976fa6c5af5b20ea8741d4f5fc3c)