Fix flaky test in insert_root_partition_truncate_deadlock
The injected fault 'func_init_plan_end' is set only triggered once, and it's
possible that the fault can be triggered by dtx recovery process before than
insert statement. So in this case, the insert statement is not blocked which
causes the pipeline failed.

Inject the fault 'func_init_plan_end' in the same session as the insert, and
specify the current sessionid, so that the fault can't be triggered by other
sessions.

The test insert_root_partition_truncate_deadlock with gdd have the same problem.