Emit ERROR if gp_fastsequence entry goes backward during update
gp_fastsequnce entry semantic is to only move forward. Hence, add
check to protect and catch if this assumption is broken. There existed
bug which is fixed now, during AO table truncate where this assumption
was broken and it was very hard to trace the RCA in absence of this
check.