Preserve param info when bringing a path to OuterQuery locus
When we bring a path to OuterQuery locus, we need to keep its param info,
because this path may further join with other paths. For instance,

	select * from a where a.i in
		(select count(b.j) from b, c,
			lateral (select * from d where d.j = c.j limit 10) s
		where s.i = a.i
		);

The path for subquery 's' requires parameter from 'c'. When we bring this path
to OuterQuery locus, its param info needs to be preserved, so that when joining
's' with 'b' we can have correct param info.