[AORO] change addition assignment to assigment
In function `GetAllFileSegInfo_pg_aoseg_rel`, when handling each
tuple, we `palloc0` a FileSegInfo, `oneseginfo` is just a pointer
to the zero-allocated memory, using `+=` with the left operand 0
is the same effect as `=`, and `+=` should burn more cpu cycles
than `=`.

I'm not sure the compiler will optimize this kind of `+=` to `=`,
even if it does, using `=` is more accurate since here it is not
a accumulate semantic.

Signed-off-by: Junwang Zhao <zhjwpku@gmail.com>
(cherry picked from commit 6805b17f4ebf3345a0b149c82d676e94dbdc06d1)