Disallow Insert and Delete triggers on SplitUpdate
Considering that the original command is UPDATE for a SplitUpdate, fire
INSERT and DELETE triggers may lead to the wrong action to be enforced.
And the triggers in GPDB may require cross segments data changes, disallow
the INSERT and DELETE triggers on a SplitUpdate.

(cherry picked from commit c33386e5b59e2032395ae701ed3434f987489ac8)