Fix an issue with rle_type when changing table from AO to AOCO
It is found that when changing from AO to AOCO we cannot specify
`compresstype=rle_type` which would complain that it is not
applicable to AO table, despite that it should be a valid usage.

This is because during setting new reloptions, we miss the
information about the new AM being AOCO, and the code behavior
is to validate the reloptions as AO tables.

Fixing it by passing the new AM into ATExecSetRelOptions() based
on which we can validate the reloptions. If there is no AM
change, an InvalidOid will be passed in and we will use the
current table AM to validate the reloptions.

(cherry picked from commit 374d05728a00a447af9099bbfc710e86008f8ea8)