Resolve a GPDB_12_MERGE_FIXME in regress/dsp.
The first FIXME is introduced by the commit from
merge branch 002f61d8. Remove it and run the full
pipelines it passes the tests.

The second FIXME is asking if it is necassary to
test GUC gp_default_storage_options should keep
consistent among master and segments. Although
gpconfig's unittest might cover this, it does no
harm to also test here for this important GUC.