Resolve fixme comment in ATExecAddColumn() about why no lock child table
In ATExecAddColumn() where we find child tables to apply a flag,
we used to open the tables w/o lock. That is because we have already
held a lock on the parent and we implicitly believed that no one
else could alter the child tables now. That is still correct, but
the upstream had introduced an expectation for relation_open()
in 4b21acf522d751ba5b6679df391d5121b6c4a35f that all callers of
relation_open() should have acquired a lock except in bootstrap
mode. I do not think it's important to still use NoLock and try
to workaround that assert in relation_open() in some way.
Therefore, resolving the FIXME as it is.

(cherry picked from commit 205c4b1150993679f6c4121d84700d56d7166470)