Resolve a FIXME in MergeAttributes()
In MergeAttributes() when we merge attributes of the parents, we
avoided checking their refcount because in GPDB, ALTER TABLE ...
ADD PARTITION or SPLIT PARTITION couldn't meet the upstream
expectation that at this point the partition root shouldn't be
in use.

But that seems to be an overkill. That check would be useful for
majority of the other cases. Therefore, now adding a new filed
in CreateStmt and passing that to MergeAttributes() so that we
only avoid doing so for ALTER TABLE ... ADD|SPLIT PARTITION.
Also make sure to set the new field to false in all the other
occurrences but it should be OK if we miss some because CreateStmt
should be palloc'ed.

(cherry picked from commit 03b31902f60577f1c389c8821244e3f3b8b30702)