Cleanup a FIXME in ATExecAddIndex
We used to always dispatch in DefineIndex() but now we do not if
it is part of the ALTER TABLE command execution. So we no longer
need to skip DefineIndex for QEs. Confirmed and remove this FIXME.

(cherry picked from commit eac1b11d50cb8fd50b7d7388d45749bb5c559e9f)