Change optimizer_array_expansion_threshold default value to 20 (#14032)
It is expensive to derive constraints and stats if we expand the large array
into CNF/DNF form. 100 is still a large default value, so lower it down to 20
to get better performance for most queries.

The value is initially set to 25, then we use that GUC to check whether
ConstArray into array expr for Orca, then we changed to 100 because if the
value is too small we may not be able to do partition elimination.  Now on
master/7X branch we always expand ConstArray into array expr for Orca,
partition elimination will not be affected by this GUC. So setting it to a
smaller value to get better performance.

Ref: GPQP-31
(cherry picked from commit 5dc8fc2c9661bf76b66cb522b99067053643dc07)