Enable direct dispatch for DML statements with ORCA (#13471)
It was observed that, in response to a DELETE based DML query wherein we
know that the data resides on a particular segment, the commands should
be issued by Orca to that segment only. However, it was found that the
commands were issued to all the segment.  This behavior was
cross-checked with the Legacy Planner, it was found that Legacy Planner
was sending the query to single segment only.

To correct this behavior changes were done in the following files:

1. FILE NAME : CTranslatorExprToDXL.cpp

In function CtranslatorExprToDXL::PdxlnDML , object for
CDXLDirectDispatchInfo is created by calling GetDXLDirectDispatchInfo
function.

Before the update, in the GetDXLDirectDispatchInfo function, null
pointer was returned for any DML command other than INSERT.

Now, this condition has been changed to include DELETE command also,
thus now for INSERT and DELETE, nullpointer shall not be returned and
object for CDXLDirectDispatchInfo class shall be created.

2. FILE NAME :: CTranslatorDXLToPlStmt.cpp

The above created object is checked while creating the planned statement
for the query to enable the direct dispatch flag in the planned
statment.

Earlier this was allowed for INSERT command only, changes were done to
enable this logic for DELETE command also.

(cherry picked from commit 26c35a8ab16ffd54a20ed05140fccbd95bd7f550)