Resolve a FIXME in merge_leaf_stats()
A FIXME was left noting whether to use AccessShareLock or NoLock
when opening child tables to merge their stats. Since we have
acquired ShareUpdateExclusiveLock on the parent table when we
are at merge_leaf_stats(), we don't need extra lock to guard
against concurrent DROP of either the parent or the child (which
requries AccessExclusiveLock on the parent).
Concurrent UPDATE is possible but because we are not updating the table
ourselves, NoLock is sufficient here.

(cherry picked from commit 96d06cafad8f6ef573a705d21941370618685fba)