Exclude AMs with internal handlers in gpcheckcat dependency checks
Creating AMs with internal handlers will not have entries in `pg_depend`,
which is reasonable since internal handlers could not be dropped, but
`gpcheckcat` is not happy with it.

This commit excludes that and fixes the `getName()` deprecated issue.

(cherry picked from commit f3ea5c6c5b86202b50591327bda03c94561296ca)