Add support for Foreign Scans in Orca
This adds support for performing simple scans on Foreign tables in Orca.
This does predicate pushdown and column projection if the FDWs support
these optimizations, but it doesn't do more complex optimizations such
as agg pushdown/remote joins.

There are 3 FDW functions that we call: GetForeignRelSize,
GetForeignPaths, and GetForeignPlan. As far as I can tell, this is where
the fdw_private field can be modified/populated. Note that we don't care
about the result of these function calls-- it's strictly to populate
fdw_private.

The tricky part of this commit was replicating some of the structures
needed for the FDW api function calls. Since we have to call these
functions, we needed to create "dummy" structures that still had the
information we care about.

Support for scanning foreign partitions will be added in a subsequent
commit.

(cherry picked from commit 08702ccc5f479e17e5a6c28ed955e72cd6ed7500)