gp_replica_check: resolve FIXMEs
FIXMEs were related to adding support for checking dynamic access
methods. Given we are far away from supporting all built-in access
methods, dynamic support is not on the radar at all. Plus, having a
new access method is not a common act at all. So, we can cross the
bridge of converting RelationTypeData to a dynamic array instead of
the static array when we get there. Hence, for now, resolving the
FIXME by manually updating the static array to reflect all the current
access methods.

Note: with this change now "ao" gets split into ao_row and
ao_column. This is nice as can run more granular checking if desired.

Also, converted GPDB_94_MERGE_FIXME to GPDB_FEATURE_NOT_SUPPORTED in
Makefile. As we are yet to test/validate/enable this tool for various
index access methods - which we should do separately.