Enable direct dispatch if distribution column is of 'varchar' type and update hash function selection based on distribution policy of table
ENABLE DIRECT DISPATCH
ORCA did not perform direct dispatch when the distribution column was of type 'varchar'
With this update this error is fixed. For direct dispatch, in the CTranslatorExprToDXLUtils::FDirectDispatchable(,,) datum and
distribution column OIds are compared. Direct dispatch is performed, if they are equal.

Consider the following query which failed for a table (create table t1_varchar(col1_varchar varchar, col2_int int)):
   explain select * from t1_varchar where col1_varchar = ‘c’;

In this case both the Old field in IMDid class, were different, thus direct dispatch was not performed.
To fix this, it is checked if a cast exist between them and if that cast is binary coercible.

UPDATE HASH FUNCTION SELECTION
After the code was corrected for direct dispatch, a bug surfaced out in the code due the fix.
During the pipeline run, it was found that following tests in citext.sql were failed

SETUP
CREATE EXTENSION citext;
CREATE TEMP TABLE srt ( name CITEXT);
INSERT INTO srt (name) VALUES ('abb'), ('ABA'), ('ABC'), ('abd');

QUERIES
SELECT LOWER(name) as aba FROM srt WHERE name = 'ABA'::text;
SELECT LOWER(name) as aba FROM srt WHERE name = 'ABA'::varchar;
SELECT LOWER(name) as aba FROM srt WHERE name = 'ABA'::bpchar;

Before the fix for direct dispatch, these queries were running successfully as direct dispatch
was not performed for them, Planner on the other hand was performing direct dispatch for them.
With the update in direct dispatch concept, ORCA was trying to perform direct dispatch
in these queries also, but a wrong segment number was generated by the hashing function.
On debugging the code, it was found that, wrong segment number was selected due to selection of
wrong hash function. On further analysis, it was found that for direct dispatch ORCA selected hash function
based on the const typeoid. Planner on the other hand selected hash function based on the distribution
policy of the relation.To correct this bug, the hash function selection for direct dispatch was also
made on the distribution policy of the relation.

Implementation Key points
1. During the DXL to planned statement stage, we extract the information of available RTEs in FROM clause
2. We perform direct dispatch based on the distribution policy, if there is only one unique RTE
in FROM clause of query, else we go back to the old logic.

(cherry picked from commit 14a78014f5488d89fbc5f48a341f10c2798fa227)