Fix bug that Orca fails to decorrelate subqueries order by outer reference (#14905)
This patch reinstates the logic before commit 30e79d2, so that when there is
order by clause in the subquery, we can safely remove the order by clause. The
previous logic was not wrong, but will keep the outer reference as correlated
subquery, which is also worse plan but correct correlated plan. But in our
preprocessing step, 26th, it is doing wrong on decorrelating subquery, in which
case, we can't decorrelate actually.

The commit 30e79d2 added a fix that allows adding a project in the translator
that will echo the outer ref from within the subquery.

`select foo.c from foo where b in 
  (select bar.q from bar where bar.r =10 order by foo.c);`

ORCA is currently unable to decorrelate sub-queries that contain project
nodes.For the above query the parser adds foo.c to the target list of the
subquery, that would get an echo projection and thus would prevent decorelation
by ORCA as per the mentioned commit. This patch will not project a ORDER BY
column so that we can safely remove the order by clause.

Author: @DevChattopadhyay
(cherry picked from commit ff712241cd87efa03695bd980bb77607db95f010)