[ORCA] Allow push down of filter with BETWEEN predicate (#14872)
This comimit enables additional opportunities for predicate push down
based on a synergetic relationship between join conditions and select
filters.  For example, let's use query:

    SELECT count(*) FROM t2 INNER JOIN t1 ON t2.c BETWEEN t1.a AND t1.b WHERE t2.c = 2;

Input:
    +--CLogicalSelect
       |--CLogicalNAryJoin
       |  |--CLogicalGet "t2" ("t2"), Columns: ["c" (0), "d" (1),...
       |  |--CLogicalGet "t1" ("t1"), Columns: ["a" (9), "b" (10),...
       |  +--CScalarBoolOp (EboolopAnd)
       |     |--CScalarCmp (>=)
       |     |  |--CScalarIdent "c" (0)
       |     |  +--CScalarIdent "a" (9)
       |     +--CScalarCmp (<=)
       |        |--CScalarIdent "c" (0)
       |        +--CScalarIdent "b" (10)
       +--CScalarCmp (=)
          |--CScalarIdent "c" (0)
          +--CScalarConst (2)

Notice that the join predicate references columns (a, b, and c) coming
from either sides of the join. Typically this means the predicate cannot
be pushed down.  However, the select filter above the join predicate
filters on c=2.  We can utilize that in conjunction with the join
condition to filter out more rows. We do that outputing:

Output:
    +--CLogicalSelect
       +--CLogicalSelect
       |  |--CLogicalNAryJoin
       |  |  |--CLogicalGet "t2" ("t2"), Columns: ["c" (0), "d" (1),...
       |  |  |--CLogicalGet "t1" ("t1"), Columns: ["a" (9), "b" (10),...
       |  |  +--CScalarBoolOp (EboolopAnd)
       |  |     |--CScalarCmp (>=)
       |  |     |  |--CScalarIdent "c" (0)
       |  |     |  +--CScalarIdent "a" (9)
       |  |     +--CScalarCmp (<=)
       |  |        |--CScalarIdent "c" (0)
       |  |        +--CScalarIdent "b" (10)
       |  +--CScalarBoolOp (EboolopAnd)
       |     |--CScalarCmp (>=)
       |     |  +--CScalarConst (2)
       |     |  +--CScalarIdent "a" (9)
       |     +--CScalarCmp (<=)
       |     |  +--CScalarConst (2)
       |        +--CScalarIdent "b" (10)
       +--CScalarCmp (=)
          |--CScalarIdent "c" (0)
          +--CScalarConst (2)

Note: One might naively think that it is possible to directly replace
the constant identifier in the join predicate.  However, that would be a
mistake as it could prevent hash join alternative plans. This is why we
must instead add a new select filter above the join.

The preprocessor step in this commit can add new select filters that
were constructed by substituting constants into join conditions.  It is
possible that the new filter is of the form:

    (const relop ident)

However, prior to this commit the constraint interval is constructed
using a scalar compare operator that is expected to be an expression of
the form:

    (ident relop const)

There doesn't seem a good reason to not to support that form too. In
fact there's a FIXME from 2012 that suggests maybe we should enable even
more forms.  So this commit also enables the above form.

(cherry picked from commit 31254f40fc44c132c35ad4625f6fad9b3ed9a1cb)