Prevent duplicate filter issue
Following example produces a duplicate fliter:
    ```
    CREATE TABLE t1(a bigint);
    CREATE TABLE t2(a bigint);

    EXPLAIN (costs off)
    SELECT * FROM t1 INNER JOIN t2 ON t2.a = t1.a WHERE t1.a = 17353232;
                                          QUERY PLAN
    --------------------------------------------------------------------------------------
     Gather Motion 3:1  (slice1; segments: 3)
       ->  Hash Join
             Hash Cond: (t1.a = t2.a)
             ->  Seq Scan on t1
                   Filter: (a = 17353232)
             ->  Hash
                   ->  Seq Scan on t2
                         Filter: ((a OPERATOR(pg_catalog.=) 17353232) AND (a = 17353232))
     Optimizer: Pivotal Optimizer (GPORCA)
    ```

In order to understand the issue we need to look closely at the
expression tree.  After parse time the constant 173532320 is interpreted
as an int4.
    ```
    +--CLogicalSelect
       |--CLogicalNAryJoin
       |  |--CLogicalGet "t1" ("t1"), Columns: ["a" (0), "ctid" (1), ...
       |  |--CLogicalGet "t2" ("t2"), Columns: ["a" (8), "ctid" (9), ...
       |  +--CScalarCmp (=)
       |     |--CScalarIdent "a" (8)
       |     +--CScalarIdent "a" (0)
       +--CScalarCmp (=)
          |--CScalarIdent "a" (0)
          +--CScalarConst (173532320)
    ```

Notice that the top CScalarCmp operator id is 416 (int84eq) which
compares int8 to int4. While the bottom CScalarCmp operator id is 410
(int8eq) which comparees int8 to int8.
    ```
    (gdb) pp ((CScalarCmp*)(*pexprTrimmed2)[1]->m_pop)->m_mdid_op
    (0.416.1.0)
    (gdb) pp ((CScalarCmp*)(*(*pexprTrimmed2)[0])[2]->m_pop)->m_mdid_op
    (0.410.1.0)
    ```

Issue is that the new BETWEEN predicate preprocessor step will create a
new select filter with the int8eq operator used in the join condition.
Besides being wrong because of the type mismatch of CScalarConst, it
will also build a filter that is technically different than the top
int84eq operator select filter. Thus during normalization we cannot
deduplicate the two filters.

Fix in this commit is to prevent the preprocessor step from substituting
constants when types do not match. Admittedly there may be missed
opportunities for push down. However, the alternative to fix up the
CScalarCmp types and arguments is a larger scoped change. A workaround
using fix in this commit is to cast the constant to the appropriate type
in the query.

(cherry picked from commit dababb3773a122283e71dd7040a66724731770ea)