Force two-stage local aggregate to remove duplicates
In a two-stage hash aggregate plan, if the local stage aggregate hash
table fills up then it has one of two options:

 1) spill to disk
 2) stream to next stage aggreegate.

Issue is if we stream a DQA to next stage and the next stage isn't
prepared to handle duplicates, then we could encounter wrong results.
That was observed in the following plan:

    SELECT a, COUNT(DISTINCT b) FROM t GROUP BY a;

                                       QUERY PLAN
    --------------------------------------------------------------------------------
     Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..3366.93 rows=7 width=12)
       ->  HashAggregate  (cost=0.00..3366.93 rows=3 width=12)
             Group Key: a
             ->  HashAggregate  (cost=0.00..3212.44 rows=1273152 width=8)
                   Group Key: a, b
                   ->  Seq Scan on test  (cost=0.00..640.00 rows=10000034 width=8)
     Optimizer: Pivotal Optimizer (GPORCA)
    (7 rows)

The second HashAggregate used hybrid hashtable in streaming mode. That
is, when the hashtable enumerating all distinct pairs of A and B fills
up, we start it over instead of spilling to disc. This results in
duplicates and we get more tuples as a result of this query.

It seems that we are using streaming hashtable, because optimizer thinks
that it is okay to produce duplicates for this plan HashAggregate
(<dxl:Aggregate AggregationStrategy="Hashed" StreamSafe="true"> from
minidump), but it is not. The fix here simply prohibits the aggregate
node from producing duplicates, and it resolves the problem.

Authored-by: Smyatkin Maxim (with minor edits by David Kimura)
(cherry picked from commit c069dfbcb44be51d34f6f2afd5d051a11f05eaf8)